cas: rename to store
The cas package implements more than just a CAS, so this patch gives it
a suitably generic name.

Fixes #707
