+++ rkt/run_prepared.go
@@ -21,8 +21,8 @@ import (
 	"io/ioutil"
 	"log"
 
-	"github.com/coreos/rkt/cas"
 	"github.com/coreos/rkt/stage0"
+	"github.com/coreos/rkt/store"
 )
 
 const (
@@ -70,7 +70,7 @@ func runRunPrepared(args []string) (exit int) {
 		}
 	}
 
-	ds, err := cas.NewStore(globalFlags.Dir)
+	ds, err := store.NewStore(globalFlags.Dir)
 	if err != nil {
 		stderr("prepared-run: cannot open store: %v", err)
 		return 1
