+++ test
@@ -6,15 +6,15 @@
 #
 # Run tests for one package
 #
-# PKG=./cas ./test
-# PKG=cas ./test
+# PKG=./store ./test
+# PKG=store ./test
 
 # Invoke ./cover for HTML output
 COVER=${COVER:-"-cover"}
 
 source ./build
 
-TESTABLE_AND_FORMATTABLE="cas pkg/keystore pkg/lock pkg/tar rkt stage1/init"
+TESTABLE_AND_FORMATTABLE="pkg/keystore pkg/lock pkg/tar rkt stage1/init store"
 FORMATTABLE="$TESTABLE_AND_FORMATTABLE common networking stage0/run.go version"
 
 # user has not provided PKG override
